NET6.0, NET7.0 Support, Debug mode
Added support for NET6.0 and NET7.0 which was not working before.

Added a debug mode/option which you can set with:
<CompilerVisibleProperty Include="AIG_EnableDebug" />
in an ItemGroup in the .csproj file

This will generate valuable debug information such as summary of the properties detected with inject attribute, options enabled, inject symbol detection and other.

---
Downgraded LangVersion to allow better compatibility
A bit of refactor and renaming was also done.

---
Next?: Removing debug mode from releases to minimize size.