Switch to prebuilding generator
This is a major change and in the end will result in 2.0.0 version
From now on the incremental generators will go out of use, this project will use a prebuilding generators, this provides us with many benefits, some of which I'm going to name below:

1. Easy view access to files - Since the files are now written to disk they will be viewable from file explorers. This wasn't the case before
2. Integration into custom build frameworks, mainly Godot. The code now is build as if it was users, which allows the generated code to be integrated into Godot lifecycle, this wasn't possible before - This is a major difference and positive of the new approach.
3. Possible generation speed increases, has to be measured, but the new generator has been setup in such a way that it only processes changed files meaning the generator doesnt need to regenerate everything at any change or build.
4. More possibilities, the generator is more integrated into the system meaning it can edit files as well, I don't use that in the generator yet, but it's great to be able to do that.
5. Much more

In the current state the generator performs well, the first run will be slow increasing dramatically with the second run. (Ex. with 10k files first run will take around 10s and the subsequent runs around 10-20ms - The runs are further optimized to check whether the file needs to be processed at all, this check would in most cases result in speeding up the processing by skipping not needed files, however the benchmark in this example only contains files that need to be processed which makes the whole process slower not faster, the benchmark is pretty faulty, better benchmarks should appear later)

Next changes:
- Code refactoring,
- Better dependency resolution
- Better API for creating custom generators

Explanation of next changes:
1. Code refactoring - Pretty self-explanatory, the code should be refactored so its more readable, there shouldn't be any changes in the overall logic.
2. Better dependency resolution, currently the missing dependencies are searched within C# files and the project is build with manual dll file references,
I think this could be improved, but it's a hard nut to crack, manual dll files will probably prevail.
3. Better API, more clarity on what the generator does. Later on I want an API to be made for uncomplicated adding of custom generators.

This commit marks 2.0.0 Alpha.