Better missing dependencies handling
Resolving dependencies is now a lot better as it works with a structure file which is constructed at the first run and maintained next runs.
It solves an issue where the injected class was not included in the project and couldn't be referenced by a full name.

This gives us better resolution not based on file names.

NEEDS Refactor and cleanup