WIP Refactor and removal of DI Container
Refactors the code a lot so to make it cleaner to read and more maintainable.
Removes the DI Container (DryIoc) for Contexts, which in this case are easier to use and possibly improve performance.

It requires a bit more work, but as of right now the generation works.
TO Fix: Generator doesn't delete previously generated files after [Inject] is removed. If no [Inject] are found in a changed file the other generated files should get removed.